// src/api/generated/models/createStageRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aegis Backend API
 * OpenAPI spec version: 0.1.0
 */
import type { StageKind } from './stageKind.ts';

export interface CreateStageRequest {
  /**
     * @minimum 1
     * @maximum 900
     */
  duration_seconds: number;
  kind: StageKind;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
  /**
     * @minimum 1
     * @maximum 1000
     */
  radius_m: number;
}